govuk-template: Move GraphQL code into own file
