+++ apps/govuk-template/src/server/index.ts
@@ -6,37 +6,7 @@ import AppWrap from '../common/app-wrap';
 import ErrorPage from '../common/error-page';
 import PageWrap from '../common/page-wrap';
 import pageLoader from '../common/page-loader';
-
-import { makeExecutableSchema } from '@graphql-tools/schema';
-
-// Some fake data
-const books = [
-  {
-    title: "Harry Potter and the Sorcerer's stone",
-    author: 'J.K. Rowling',
-  },
-  {
-    title: 'Jurassic Park',
-    author: 'Michael Crichton',
-  },
-];
-
-// The GraphQL schema in string form
-const typeDefs = `
-type Query { books: [Book] }
-type Book { title: String, author: String }
-`;
-
-// The resolvers
-const resolvers = {
-  Query: { books: () => books },
-};
-
-// Put together a schema
-const graphQLSchema = makeExecutableSchema({
-  typeDefs,
-  resolvers,
-});
+import graphQLSchema from './graphql';
 
 const setup = () => {
   const assets = (
@@ -128,6 +98,7 @@ if (module.hot) {
   module.hot.accept([
     '@not-govuk/engine',
     './config',
+    './graphql',
     '../../dist/public/entrypoints.json',
     '../../webpack.config'
   ], restart);
